full profile is essentially integrated; mini profile next
