--- app/components/Profile/index.ts
@@ -1,3 +1,0 @@
-export * from './Profile';
-export * from './Cogno';
-export * from './transaction';
